bot_w_ia/requirements.txt
flask
requests
llama-cpp-python
